rename url shortener projects to urls service as all urls will be handled by this service
