+++ Materials/SciMaterials.Materials.Api/ResourcesRegister.cs
@@ -37,10 +37,9 @@ using SciMaterials.Services.API.Services.Files.Stores;
 using SciMaterials.Services.API.Services.Resources;
 using SciMaterials.Services.API.Services.Tags;
 using SciMaterials.Services.API.Services.Urls;
-using SciMaterials.Services.ShortLinks;
-using SciMaterials.UrlShortener.Contracts;
-using SciMaterials.UrlShortener.Contracts.Settings;
-using SciMaterials.UrlShortener.Domain;
+using SciMaterials.UrlsService.Contracts;
+using SciMaterials.UrlsService.Contracts.Settings;
+using SciMaterials.UrlsService.Domain;
 using SciMaterials.WebApi.Clients.Categories;
 using SciMaterials.WebApi.Clients.Comments;
 using SciMaterials.WebApi.Clients.ContentTypes;
