+++ SciMaterials.sln
@@ -63,13 +63,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		Directory.Build.targets = Directory.Build.targets
 	EndProjectSection
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "UrlShortener", "UrlShortener", "{4A693F95-9ECC-4DA0-9BB2-A9EA1BC4D4D7}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "UrlsService", "UrlsService", "{6E00A596-5CFB-436A-BF7E-E324C503FCBE}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SciMaterials.UrlShortener.Api", "UrlShortener\SciMaterials.UrlShortener.Api\SciMaterials.UrlShortener.Api.csproj", "{F337A2DD-CA89-4C1A-8B08-05B845AB0626}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SciMaterials.UrlsService.Api", "UrlsService\SciMaterials.UrlsService.Api\SciMaterials.UrlsService.Api.csproj", "{26F6E741-5D4D-48E2-8515-6930767AE4BF}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SciMaterials.UrlShortener.Contracts", "UrlShortener\SciMaterials.UrlShortener.Contracts\SciMaterials.UrlShortener.Contracts.csproj", "{F4BE2912-8EA4-43E1-B9FB-3C4AF8FEE7F0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SciMaterials.UrlsService.Contracts", "UrlsService\SciMaterials.UrlsService.Contracts\SciMaterials.UrlsService.Contracts.csproj", "{347FEC80-DFE0-421A-BA11-08CAE72A24F9}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SciMaterials.UrlShortener.Domain", "UrlShortener\SciMaterials.UrlShortener.Domain\SciMaterials.UrlShortener.Domain.csproj", "{9C4B92CE-CC22-41B0-A59F-F86DB52AAFC9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SciMaterials.UrlsService.Domain", "UrlsService\SciMaterials.UrlsService.Domain\SciMaterials.UrlsService.Domain.csproj", "{40DCE8AA-8B19-40CB-A9D5-51CC01CBFCC8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -157,18 +157,18 @@ Global
 		{8A2B273F-0BD5-49DC-9416-2389056CD3A8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8A2B273F-0BD5-49DC-9416-2389056CD3A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8A2B273F-0BD5-49DC-9416-2389056CD3A8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F337A2DD-CA89-4C1A-8B08-05B845AB0626}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F337A2DD-CA89-4C1A-8B08-05B845AB0626}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F337A2DD-CA89-4C1A-8B08-05B845AB0626}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F337A2DD-CA89-4C1A-8B08-05B845AB0626}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F4BE2912-8EA4-43E1-B9FB-3C4AF8FEE7F0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F4BE2912-8EA4-43E1-B9FB-3C4AF8FEE7F0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F4BE2912-8EA4-43E1-B9FB-3C4AF8FEE7F0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F4BE2912-8EA4-43E1-B9FB-3C4AF8FEE7F0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9C4B92CE-CC22-41B0-A59F-F86DB52AAFC9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9C4B92CE-CC22-41B0-A59F-F86DB52AAFC9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9C4B92CE-CC22-41B0-A59F-F86DB52AAFC9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9C4B92CE-CC22-41B0-A59F-F86DB52AAFC9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{26F6E741-5D4D-48E2-8515-6930767AE4BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{26F6E741-5D4D-48E2-8515-6930767AE4BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{26F6E741-5D4D-48E2-8515-6930767AE4BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{26F6E741-5D4D-48E2-8515-6930767AE4BF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{347FEC80-DFE0-421A-BA11-08CAE72A24F9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{347FEC80-DFE0-421A-BA11-08CAE72A24F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{347FEC80-DFE0-421A-BA11-08CAE72A24F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{347FEC80-DFE0-421A-BA11-08CAE72A24F9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{40DCE8AA-8B19-40CB-A9D5-51CC01CBFCC8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{40DCE8AA-8B19-40CB-A9D5-51CC01CBFCC8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{40DCE8AA-8B19-40CB-A9D5-51CC01CBFCC8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{40DCE8AA-8B19-40CB-A9D5-51CC01CBFCC8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -197,9 +197,9 @@ Global
 		{75EE60FB-2813-4C67-962B-77829DC59596} = {D5E202AB-8A5B-4C21-BCEB-204725886277}
 		{6C4DB170-F3D8-4386-8E23-CE292546AE25} = {F90B51A6-59E4-4026-96D7-76A0BECC2BE6}
 		{8A2B273F-0BD5-49DC-9416-2389056CD3A8} = {2628C1F7-3DCD-4B27-8B05-9D4B649E7ABC}
-		{F337A2DD-CA89-4C1A-8B08-05B845AB0626} = {4A693F95-9ECC-4DA0-9BB2-A9EA1BC4D4D7}
-		{F4BE2912-8EA4-43E1-B9FB-3C4AF8FEE7F0} = {4A693F95-9ECC-4DA0-9BB2-A9EA1BC4D4D7}
-		{9C4B92CE-CC22-41B0-A59F-F86DB52AAFC9} = {4A693F95-9ECC-4DA0-9BB2-A9EA1BC4D4D7}
+		{26F6E741-5D4D-48E2-8515-6930767AE4BF} = {6E00A596-5CFB-436A-BF7E-E324C503FCBE}
+		{347FEC80-DFE0-421A-BA11-08CAE72A24F9} = {6E00A596-5CFB-436A-BF7E-E324C503FCBE}
+		{40DCE8AA-8B19-40CB-A9D5-51CC01CBFCC8} = {6E00A596-5CFB-436A-BF7E-E324C503FCBE}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {69CA7005-2892-4D9A-9575-8BCEC07E43F2}
