+++ UrlsService/SciMaterials.UrlsService.Contracts/IUrlShortenerClient.cs
@@ -1,6 +1,6 @@
 ﻿using SciMaterials.Contracts.Result;
 
-namespace SciMaterials.UrlShortener.Contracts;
+namespace SciMaterials.UrlsService.Contracts;
 public interface IUrlShortenerClient
 {
 	string ShortenedUrlBase { get; }
