+++ UrlsService/SciMaterials.UrlsService.Domain/SciMaterials.UrlsService.Domain.csproj
@@ -5,6 +5,6 @@
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="7.0.0" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\SciMaterials.UrlShortener.Contracts\SciMaterials.UrlShortener.Contracts.csproj" />
+    <ProjectReference Include="..\SciMaterials.UrlsService.Contracts\SciMaterials.UrlsService.Contracts.csproj" />
   </ItemGroup>
 </Project>
